:recycle: :factory: Core Input System
