+++ Disappear/Assets/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e1274190b2230a41a1501c13afab868
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
