+++ Disappear/Assets/Core/Core.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fe16aff440d71e944bada5b811aa591b
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
