+++ Disappear/Assets/Core/Input System/Controls.inputactions
@@ -222,7 +222,7 @@
                     "name": "",
                     "id": "1e8cf369-5e3a-4694-8c25-b2397438245a",
                     "path": "<Mouse>/press",
-                    "interactions": "Hold",
+                    "interactions": "",
                     "processors": "",
                     "groups": "",
                     "action": "Catch",
