+++ Disappear/Assets/Health/ScriptableObjects/DyingHeathSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f9102ce2602e1344a3e0c998fa9b340
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
