+++ Disappear/Assets/Health/ScriptableObjects/HealthyHeathSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a6082ecf9822614f883b57a5e89c419
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
