+++ Disappear/Assets/Health/ScriptableObjects/HeathSO.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace WAG.Health
 {
-    [CreateAssetMenu(fileName = "New HeathSO", menuName = "SO/HeathSO")]
+    [CreateAssetMenu(fileName = "HeathSO", menuName = "SO/HeathSO")]
     public class HeathSO : ScriptableObject
     {
         [field:SerializeField] public HeathStatus Status { get; private set; }
