+++ Disappear/Assets/Health/ScriptableObjects/WoundedHeathSO.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7bd4b4fb88dcc5345b52a4809dd63fb7, type: 3}
+  m_Name: WoundedHeathSO
+  m_EditorClassIdentifier: 
+  <Status>k__BackingField: 1
+  <SpeedModifier>k__BackingField: -0.2
