+++ Disappear/Assets/Health/ScriptableObjects/WoundedHeathSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c1f6e8dfe3f431a4fbc587f49ff296d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
