+++ Disappear/Assets/Health/Scripts/HealthTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c84842702da3df4abc539873f7ae2f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
