+++ Disappear/Assets/PlayerControllers/Scripts/CameraController.cs
@@ -1,8 +1,5 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-
+using WAG.Core.Controls;
 
 public class CameraController : MonoBehaviour
 {
