+++ Disappear/Assets/PlayerControllers/Scripts/StaminaController.cs
@@ -1,7 +1,5 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
+using WAG.Core.Controls;
 
 namespace WaG
 {
