+++ Disappear/Assets/UI/Scripts/StaminaGaugeManager.cs
@@ -1,6 +1,6 @@
-using System;
+
 using System.Collections;
-using System.Collections.Generic;
+using WAG.Core.Controls;
 using UnityEngine;
 using UnityEngine.UI;
